Divided classes into packages, up button in Buy, Sell, Wishlist. Wishlist not showing its own title, instead showing app's title.
