+++ src/main/java/com/naughtybitch/discogsapi/RetrofitClient.java
@@ -17,37 +17,33 @@ public class RetrofitClient {
     anywhere in the application
      */
 
-    public static synchronized Retrofit getRetrofitClient(OkHttpClient client) {
+    public static Retrofit getRetrofitClient(OkHttpClient client) {
         // If condition to ensure we don't create multiple retrofit instances in a single application
         if (retrofit == null) {
-            synchronized (RetrofitClient.class) {
-                if (retrofit == null) {
-                    Log.i("instance", "Creating new instance");
+            Log.i("instance", "Creating new instance");
 
-                    // Defining the Retrofit using Builder
-                    retrofit = new Retrofit.Builder().baseUrl(BASE_URL) // This is the only mandatory call on Builder object.
-                            .addConverterFactory(GsonConverterFactory.create()) // Converter library used to convert response into POJO
-                            .client(client).build();
-                }
-            }
+            // Defining the Retrofit using Builder
+            retrofit = new Retrofit.Builder().baseUrl(BASE_URL) // This is the only mandatory call on Builder object.
+                    .addConverterFactory(GsonConverterFactory.create()) // Converter library used to convert response into POJO
+                    .client(client).build();
         }
         return retrofit;
     }
 
-    public static synchronized Retrofit getRetrofitClient() {
+
+    public static Retrofit getRetrofitClient() {
         // If condition to ensure we don't create multiple retrofit instances in a single application
         if (retrofit == null) {
-            synchronized (RetrofitClient.class) {
-                if (retrofit == null) {
-                    Log.i("instance", "Creating new instance");
+            Log.i("instance", "Creating new instance");
 
-                    // Defining the Retrofit using Builder
-                    retrofit = new Retrofit.Builder().baseUrl(BASE_URL) // This is the only mandatory call on Builder object.
-                            .addConverterFactory(GsonConverterFactory.create()) // Converter library used to convert response into POJO
-                            .build();
-                }
-            }
+            // Defining the Retrofit using Builder
+            retrofit = new Retrofit.Builder().baseUrl(BASE_URL) // This is the only mandatory call on Builder object.
+                    .addConverterFactory(GsonConverterFactory.create()) // Converter library used to convert response into POJO
+                    .build();
         }
         return retrofit;
     }
 }
+
+
+
