+++ src/main/java/com/naughtybitch/discogsclient/SearchableActivity.java
@@ -3,9 +3,10 @@ package com.naughtybitch.discogsclient;
 import android.app.SearchManager;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.View;
+import android.widget.ProgressBar;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -16,6 +17,8 @@ import com.naughtybitch.discogsapi.RetrofitClient;
 import com.naughtybitch.recyclerview.ResultsAdapter;
 
 import java.io.IOException;
+import java.sql.Timestamp;
+import java.util.ArrayList;
 import java.util.List;
 
 import okhttp3.Interceptor;
@@ -28,20 +31,14 @@ import retrofit2.Retrofit;
 
 public class SearchableActivity extends AppCompatActivity {
 
-    private static SearchableActivity instance;
+    Context context = this;
     List<Result> results;
     RecyclerView recyclerView;
+    ResultsAdapter adapter;
+    ProgressBar progressBar;
 
-    public static SearchableActivity getInstance() {
-        if (instance == null) {
-            synchronized (SearchableActivity.class) {
-                if (instance == null) {
-                    Log.i("instance", "Creating new instance");
-                    instance = new SearchableActivity();
-                }
-            }
-        }
-        return instance;
+    private static void glide() {
+
     }
 
     @Override
@@ -50,38 +47,46 @@ public class SearchableActivity extends AppCompatActivity {
         setContentView(R.layout.activity_searchable);
 
         // Lookup the recyclerView in activity layout
-//        recyclerView = (RecyclerView) findViewById(R.id.rc_view);
+
         // Initialize results
-//        results = Result.createResultsList(1);
-//        // Create adapter passing in the sample result data
-//        adapter = new ResultsAdapter(results);
-//        // Attach the adapter to the recyclerView populate items
-//        recyclerView.setAdapter(adapter);
-//        // Set layout manager to position the items
-//        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+
+        // Create adapter passing in the sample result data
+
+        // Attach the adapter to the recyclerView populate items
+
+        // Set layout manager to position the items
+
         // That's all!
         initViews();
         Intent intent = getIntent();
         if (Intent.ACTION_SEARCH.equals(intent.getAction())) {
             String query = intent.getStringExtra(SearchManager.QUERY);
+            Log.i("query", query);
             searchQuery(query);
         }
     }
 
     private void initViews() {
+        progressBar = (ProgressBar) findViewById(R.id.progress_circular);
         recyclerView = (RecyclerView) findViewById(R.id.rc_view);
+        adapter = new ResultsAdapter();
+        recyclerView.setAdapter(adapter);
         recyclerView.setLayoutManager(new LinearLayoutManager(SearchableActivity.this));
     }
 
     public void searchQuery(String query) {
-        SharedPreferences userPreferences = getSharedPreferences("userPreferences", Context.MODE_PRIVATE);
-        String access_token = userPreferences.getString("access_token", "null");
-        String access_token_secret = userPreferences.getString("access_token_secret", "null");
         DiscogsClient instance = DiscogsClient.getInstance();
+        Timestamp currentTimestamp = instance.currentTimeStamp();
+        ArrayList<String> token = instance.getCredentials(context);
         final String auth = "OAuth oauth_consumer_key=\"" + instance.getConsumer_key() + "\", " +
-                "oauth_token=\"" + access_token + "\", " +
-                "oauth_signature=\"" + instance.getConsumer_secret() + access_token_secret + "\", " +
-                "oauth_signature_method=\"PLAINTEXT\"";
+                "oauth_nonce=\"" + currentTimestamp.getTime() + "\", " +
+                "oauth_token=\"" + token.get(0) + "\", " +
+                "oauth_signature_method=\"PLAINTEXT\", " +
+                "oauth_timestamp=\"" + currentTimestamp.getTime() + "\", " +
+                "oauth_version=\"1.0\", " +
+                "oauth_signature=\"" + instance.getConsumer_secret() + token.get(1) + "\"";
+
+        Log.i("header", auth);
 
         // Define the interceptor, add authentication headers
         Interceptor interceptor = new Interceptor() {
@@ -89,6 +94,7 @@ public class SearchableActivity extends AppCompatActivity {
             public okhttp3.Response intercept(Chain chain) throws IOException {
                 Request.Builder onGoing = chain.request().newBuilder();
                 onGoing.addHeader("Authorization", auth);
+                onGoing.addHeader("User-Agent", "Discogsnect/0.1 +http://discogsnect.com");
                 return chain.proceed(onGoing.build());
             }
         };
@@ -113,7 +119,6 @@ public class SearchableActivity extends AppCompatActivity {
          */
         Call<SearchResponse> call = discogsAPI.getSearchResult(query, 50, 1);
         /*
-
         This is the line which actually sends a network request. Calling enqueue() executes a call
         asynchronously. It has two callback listeners which will invoked on the main thread
          */
@@ -123,10 +128,12 @@ public class SearchableActivity extends AppCompatActivity {
                 /* This is the success callback. Though the response type is JSON, with Retrofit
                 we get the response in the form of SearchResponse POJO class
                  */
+                Log.i("response", "Status " + response.code());
                 if (response.body() != null) {
+                    progressBar.setVisibility(View.GONE);
                     SearchResponse sResponse = response.body();
                     results = sResponse.getResults();
-                    ResultsAdapter adapter = new ResultsAdapter(results);
+                    adapter = new ResultsAdapter(context, results);
                     recyclerView.setAdapter(adapter);
                 }
             }
