+++ src/main/java/com/naughtybitch/discogsclient/album/ImageAdapter.java
@@ -1,4 +1,4 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.album;
 
 import android.content.Context;
 import android.view.View;
@@ -7,6 +7,8 @@ import android.widget.ImageView;
 
 import androidx.annotation.NonNull;
 import androidx.viewpager.widget.PagerAdapter;
+
+import com.naughtybitch.discogsclient.R;
 
 
 public class ImageAdapter extends PagerAdapter {
