+++ src/main/java/com/naughtybitch/discogsclient/buy/BuyMusicActivity.java
@@ -1,4 +1,12 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.buy;
+
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
@@ -11,20 +19,22 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.widget.Toast;
+import com.google.android.material.appbar.CollapsingToolbarLayout;
 import com.google.android.material.navigation.NavigationView;
+import com.naughtybitch.discogsclient.explore.ExploreActivity;
+import com.naughtybitch.discogsclient.MainActivity;
+import com.naughtybitch.discogsclient.profile.ProfileActivity;
+import com.naughtybitch.discogsclient.R;
+import com.naughtybitch.discogsclient.sell.SellMusicActivity;
+import com.naughtybitch.discogsclient.settings.SettingsActivity;
+import com.naughtybitch.discogsclient.wishlist.WishlistActivity;
 
-public class SellMusicActivity extends AppCompatActivity implements
-        SellMusicFragment.OnFragmentInteractionListener,
-        OrderFragment.OnFragmentInteractionListener,
-        OfferFragment.OnFragmentInteractionListener,
-        InventoryFragment.OnFragmentInteractionListener{
+public class BuyMusicActivity extends AppCompatActivity implements
+        BuyMusicFragment.OnFragmentInteractionListener,
+        CartFragment.OnFragmentInteractionListener,
+        DWFragment.OnFragmentInteractionListener,
+        PurchasesFragment.OnFragmentInteractionListener,
+        OIMFragment.OnFragmentInteractionListener {
 
     private DrawerLayout drawerLayout;
     private ActionBarDrawerToggle toggle;
@@ -33,10 +43,12 @@ public class SellMusicActivity extends AppCompatActivity implements
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sell_music);
+        setContentView(R.layout.activity_buy_music);
 
         // Custom ActionBar
         final Toolbar myToolbar = (Toolbar) findViewById(R.id.my_toolbar);
+        final CollapsingToolbarLayout collapsingToolbarLayout = (CollapsingToolbarLayout) findViewById(R.id.collapse_toolbar_layout);
+        collapsingToolbarLayout.setTitleEnabled(false);
         setSupportActionBar(myToolbar);
 
         // DrawerLayout
@@ -48,11 +60,25 @@ public class SellMusicActivity extends AppCompatActivity implements
         if (actionBar != null) {
             actionBar.setHomeButtonEnabled(true);
             actionBar.setDisplayHomeAsUpEnabled(true);
-            actionBar.setDisplayShowTitleEnabled(false);
         }
+        getSupportFragmentManager().addOnBackStackChangedListener(
+                new FragmentManager.OnBackStackChangedListener() {
+                    @Override
+                    public void onBackStackChanged() {
+                        if (getSupportFragmentManager().getBackStackEntryCount() == 0) {
+                            toggle.setDrawerIndicatorEnabled(true);
+                            setTitle(R.string.music_buy);
+                        }
+                        if (getSupportFragmentManager().getBackStackEntryCount() > 0) {
+                            toggle.setDrawerIndicatorEnabled(false);
+                            getSupportActionBar().setHomeButtonEnabled(true);
+                            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+                        }
+                    }
+                });
 
         // Create a new Fragment to be placed in the activity layout
-        SellMusicFragment firstFragment = SellMusicFragment.newInstance();
+        BuyMusicFragment firstFragment = BuyMusicFragment.newInstance();
 
         // In case this activity was started with special instructions from an
         // Intent, pass the Intent's extras to the fragment as arguments
@@ -67,37 +93,37 @@ public class SellMusicActivity extends AppCompatActivity implements
     public void navigationViewHandler() {
         navigationView = (NavigationView) findViewById(R.id.navigation_view);
         navigationView.setItemIconTintList(null);
-        navigationView.setCheckedItem(R.id.sell_music);
+        navigationView.setCheckedItem(R.id.buy_music);
         navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
                 switch (menuItem.getItemId()) {
                     case R.id.home:
-                        startActivity(new Intent(SellMusicActivity.this, MainActivity.class));
+                        startActivity(new Intent(BuyMusicActivity.this, MainActivity.class));
                         break;
                     case R.id.profile:
-                        startActivity(new Intent(SellMusicActivity.this, ProfileActivity.class));
-                        Toast.makeText(SellMusicActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(BuyMusicActivity.this, ProfileActivity.class));
+                        Toast.makeText(BuyMusicActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.wish_list:
-                        startActivity(new Intent(SellMusicActivity.this, WishlistActivity.class));
-                        Toast.makeText(SellMusicActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(BuyMusicActivity.this, WishlistActivity.class));
+                        Toast.makeText(BuyMusicActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.buy_music:
-                        startActivity(new Intent(SellMusicActivity.this, BuyMusicActivity.class));
-                        Toast.makeText(SellMusicActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
+                        drawerLayout.closeDrawer(GravityCompat.START);
+                        Toast.makeText(BuyMusicActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.sell_music:
-                        drawerLayout.closeDrawer(GravityCompat.START);
-                        Toast.makeText(SellMusicActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(BuyMusicActivity.this, SellMusicActivity.class));
+                        Toast.makeText(BuyMusicActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.settings:
-                        startActivity(new Intent(SellMusicActivity.this, SettingsActivity.class));
-                        Toast.makeText(SellMusicActivity.this, "SettingsActivity", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(BuyMusicActivity.this, SettingsActivity.class));
+                        Toast.makeText(BuyMusicActivity.this, "SettingsActivity", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.explore:
-                        startActivity(new Intent(SellMusicActivity.this, ExploreActivity.class));
-                        Toast.makeText(SellMusicActivity.this, "ExploreActivity", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(BuyMusicActivity.this, ExploreActivity.class));
+                        Toast.makeText(BuyMusicActivity.this, "ExploreActivity", Toast.LENGTH_SHORT).show();
                         break;
                     default:
                         return true;
@@ -131,8 +157,15 @@ public class SellMusicActivity extends AppCompatActivity implements
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
 
         switch (item.getItemId()) {
+            case android.R.id.home:
+                FragmentManager fm = getSupportFragmentManager();
+                if (fm.getBackStackEntryCount() > 0) {
+                    fm.popBackStack();
+                    return true;
+                }
+                break;
             case R.id.search:
-                Intent intent = new Intent(SellMusicActivity.this, ExploreActivity.class);
+                Intent intent = new Intent(BuyMusicActivity.this, ExploreActivity.class);
                 startActivity(intent);
                 break;
         }
