+++ src/main/java/com/naughtybitch/discogsclient/buy/BuyMusicFragment.java
@@ -1,4 +1,4 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.buy;
 
 import android.content.Context;
 import android.net.Uri;
@@ -14,20 +14,24 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.TextView;
+import android.widget.Toast;
+
+import com.naughtybitch.discogsclient.R;
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link CartFragment.OnFragmentInteractionListener} interface
+ * {@link BuyMusicFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link CartFragment#newInstance} factory method to
+ * Use the {@link BuyMusicFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class CartFragment extends Fragment
-        //implements
-        //View.OnClickListener
-        {
+public class BuyMusicFragment extends Fragment implements
+        View.OnClickListener,
+        CartFragment.OnFragmentInteractionListener,
+        DWFragment.OnFragmentInteractionListener,
+        PurchasesFragment.OnFragmentInteractionListener,
+        OIMFragment.OnFragmentInteractionListener {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -39,32 +43,53 @@ public class CartFragment extends Fragment
 
     private OnFragmentInteractionListener mListener;
 
-    public CartFragment() {
+    public BuyMusicFragment() {
         // Required empty public constructor
     }
 
-//    private void buttonOnClickListener(View v) {
-//        Button btn_dw = v.findViewById(R.id.cart_dw_button);
-//        btn_dw.setOnClickListener(this);
-//    }
-//
-//    @Override
-//    public void onClick(View v) {
-//        switch (v.getId()){
-//            case R.id.cart_dw_button:
-//                navigateToFragment(DWFragment.newInstance());
-//                break;
-//        }
-//    }
-//
-//    private void navigateToFragment(Fragment fragment) {
-//        FragmentManager fm = getActivity().getSupportFragmentManager();
-//        FragmentTransaction fragmentTransaction = fm.beginTransaction();
-//        fragmentTransaction.add(R.id.fragment_container, fragment);
-//        fragmentTransaction.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
-//        fragmentTransaction.addToBackStack(null);
-//        fragmentTransaction.commit();
-//    }
+    private void buttonOnClickListener(View v) {
+        Button btn_cart = v.findViewById(R.id.cart);
+        Button btn_purch = v.findViewById(R.id.purchases);
+        Button btn_dw = v.findViewById(R.id.detailed_wishlist);
+        Button btn_oim = v.findViewById(R.id.offer_i_made);
+        btn_cart.setOnClickListener(this);
+        btn_purch.setOnClickListener(this);
+        btn_dw.setOnClickListener(this);
+        btn_oim.setOnClickListener(this);
+    }
+
+    @Override
+    public void onFragmentInteraction(Uri uri) {
+        // Do smt
+    }
+
+    public void onClick(View v) {
+        switch (v.getId()) {
+            case R.id.cart:
+                Toast.makeText(getActivity(), "????????", Toast.LENGTH_SHORT).show();
+                navigateToFragment(CartFragment.newInstance());
+                break;
+            case R.id.purchases:
+                navigateToFragment(PurchasesFragment.newInstance());
+                break;
+            case R.id.offer_i_made:
+                navigateToFragment(OIMFragment.newInstance());
+                break;
+            case R.id.detailed_wishlist:
+                navigateToFragment(DWFragment.newInstance());
+                break;
+        }
+    }
+
+    private void navigateToFragment(Fragment fragment) {
+        FragmentManager fm = getActivity().getSupportFragmentManager();
+        FragmentTransaction fragmentTransaction = fm.beginTransaction();
+        fragmentTransaction.replace(R.id.fragment_container, fragment);
+        fragmentTransaction.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
+        fragmentTransaction.addToBackStack(null);
+        fragmentTransaction.commit();
+    }
+
 
     /**
      * Use this factory method to create a new instance of
@@ -72,11 +97,11 @@ public class CartFragment extends Fragment
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment CartFragment.
+     * @return A new instance of fragment BuyMusicFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static CartFragment newInstance(String param1, String param2) {
-        CartFragment fragment = new CartFragment();
+    public static BuyMusicFragment newInstance(String param1, String param2) {
+        BuyMusicFragment fragment = new BuyMusicFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -84,9 +109,9 @@ public class CartFragment extends Fragment
         return fragment;
     }
 
-    public static CartFragment newInstance() {
+    public static BuyMusicFragment newInstance() {
         Bundle args = new Bundle();
-        CartFragment fragment = new CartFragment();
+        BuyMusicFragment fragment = new BuyMusicFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -104,8 +129,8 @@ public class CartFragment extends Fragment
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.fragment_cart, container, false);
-//        buttonOnClickListener(v);
+        View v = inflater.inflate(R.layout.fragment_buy_music, container, false);
+        buttonOnClickListener(v);
         return v;
     }
