+++ src/main/java/com/naughtybitch/discogsclient/buy/CartFragment.java
@@ -1,25 +1,33 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.buy;
 
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 
+import com.naughtybitch.discogsclient.R;
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link OnFragmentInteractionListener} interface
+ * {@link CartFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link ExploreFragment#newInstance} factory method to
+ * Use the {@link CartFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class ExploreFragment extends Fragment {
+public class CartFragment extends Fragment
+        implements
+        View.OnClickListener {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -31,8 +39,27 @@ public class ExploreFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public ExploreFragment() {
+    public CartFragment() {
         // Required empty public constructor
+    }
+
+    private void buttonOnClickListener(View v) {
+        Button btn_dw = v.findViewById(R.id.cart_dw_button);
+        btn_dw.setOnClickListener(this);
+    }
+
+    @Override
+    public void onClick(View v) {
+        navigateToFragment(DWFragment.newInstance());
+    }
+
+    private void navigateToFragment(Fragment fragment) {
+        FragmentManager fm = getActivity().getSupportFragmentManager();
+        FragmentTransaction fragmentTransaction = fm.beginTransaction();
+        fragmentTransaction.replace(R.id.fragment_container, fragment);
+        fragmentTransaction.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
+        fragmentTransaction.addToBackStack(null);
+        fragmentTransaction.commit();
     }
 
     /**
@@ -41,14 +68,21 @@ public class ExploreFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment ExploreFragment.
+     * @return A new instance of fragment CartFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static ExploreFragment newInstance(String param1, String param2) {
-        ExploreFragment fragment = new ExploreFragment();
+    public static CartFragment newInstance(String param1, String param2) {
+        CartFragment fragment = new CartFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    public static CartFragment newInstance() {
+        Bundle args = new Bundle();
+        CartFragment fragment = new CartFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -62,18 +96,13 @@ public class ExploreFragment extends Fragment {
         }
     }
 
-    public static ExploreFragment newInstance() {
-        Bundle args = new Bundle();
-        ExploreFragment fragment = new ExploreFragment();
-        fragment.setArguments(args);
-        return fragment;
-    }
-
+    @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_explore, container, false);
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        super.onCreateView(inflater, container, savedInstanceState);
+        View v = inflater.inflate(R.layout.fragment_cart, container, false);
+        buttonOnClickListener(v);
+        return v;
     }
 
     // TODO: Rename method, update argument and hook method into UI event
