+++ src/main/java/com/naughtybitch/discogsclient/explore/ExploreFragment.java
@@ -1,26 +1,27 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.explore;
 
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.naughtybitch.discogsclient.R;
+
+
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link OrderFragment.OnFragmentInteractionListener} interface
+ * {@link OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link OrderFragment#newInstance} factory method to
+ * Use the {@link ExploreFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class OrderFragment extends Fragment {
+public class ExploreFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -32,15 +33,8 @@ public class OrderFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public OrderFragment() {
+    public ExploreFragment() {
         // Required empty public constructor
-    }
-
-    public static OrderFragment newInstance() {
-        Bundle args = new Bundle();
-        OrderFragment fragment = new OrderFragment();
-        fragment.setArguments(args);
-        return fragment;
     }
 
     /**
@@ -49,11 +43,11 @@ public class OrderFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment OrderFragment.
+     * @return A new instance of fragment ExploreFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static OrderFragment newInstance(String param1, String param2) {
-        OrderFragment fragment = new OrderFragment();
+    public static ExploreFragment newInstance(String param1, String param2) {
+        ExploreFragment fragment = new ExploreFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -70,12 +64,18 @@ public class OrderFragment extends Fragment {
         }
     }
 
-    @Nullable
+    public static ExploreFragment newInstance() {
+        Bundle args = new Bundle();
+        ExploreFragment fragment = new ExploreFragment();
+        fragment.setArguments(args);
+        return fragment;
+    }
+
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.fragment_order, container, false);
-        return v;
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_explore, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
