+++ src/main/java/com/naughtybitch/discogsclient/sell/OfferFragment.java
@@ -1,4 +1,4 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.sell;
 
 import android.content.Context;
 import android.net.Uri;
@@ -11,17 +11,18 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+
+import com.naughtybitch.discogsclient.R;
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link DWFragment.OnFragmentInteractionListener} interface
+ * {@link OfferFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link DWFragment#newInstance} factory method to
+ * Use the {@link OfferFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class DWFragment extends Fragment {
+public class OfferFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -33,8 +34,15 @@ public class DWFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public DWFragment() {
+    public OfferFragment() {
         // Required empty public constructor
+    }
+
+    public static OfferFragment newInstance() {
+        Bundle args = new Bundle();
+        OfferFragment fragment = new OfferFragment();
+        fragment.setArguments(args);
+        return fragment;
     }
 
     /**
@@ -43,21 +51,14 @@ public class DWFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment DWFragment.
+     * @return A new instance of fragment OfferFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static DWFragment newInstance(String param1, String param2) {
-        DWFragment fragment = new DWFragment();
+    public static OfferFragment newInstance(String param1, String param2) {
+        OfferFragment fragment = new OfferFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public static DWFragment newInstance() {
-        Bundle args = new Bundle();
-        DWFragment fragment = new DWFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -75,7 +76,7 @@ public class DWFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.fragment_dw, container, false);
+        View v = inflater.inflate(R.layout.fragment_offer, container, false);
         return v;
     }
