+++ src/main/java/com/naughtybitch/discogsclient/sell/OrderFragment.java
@@ -1,4 +1,4 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.sell;
 
 import android.content.Context;
 import android.net.Uri;
@@ -11,17 +11,18 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+
+import com.naughtybitch.discogsclient.R;
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link OIMFragment.OnFragmentInteractionListener} interface
+ * {@link OrderFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link OIMFragment#newInstance} factory method to
+ * Use the {@link OrderFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class OIMFragment extends Fragment {
+public class OrderFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -33,8 +34,15 @@ public class OIMFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public OIMFragment() {
+    public OrderFragment() {
         // Required empty public constructor
+    }
+
+    public static OrderFragment newInstance() {
+        Bundle args = new Bundle();
+        OrderFragment fragment = new OrderFragment();
+        fragment.setArguments(args);
+        return fragment;
     }
 
     /**
@@ -43,21 +51,14 @@ public class OIMFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment OIMFragment.
+     * @return A new instance of fragment OrderFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static OIMFragment newInstance(String param1, String param2) {
-        OIMFragment fragment = new OIMFragment();
+    public static OrderFragment newInstance(String param1, String param2) {
+        OrderFragment fragment = new OrderFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public static OIMFragment newInstance() {
-        Bundle args = new Bundle();
-        OIMFragment fragment = new OIMFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -75,7 +76,7 @@ public class OIMFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.fragment_oim, container, false);
+        View v = inflater.inflate(R.layout.fragment_order, container, false);
         return v;
     }
