+++ src/main/java/com/naughtybitch/discogsclient/sell/SellMusicActivity.java
@@ -1,12 +1,4 @@
-package com.naughtybitch.discogsclient;
-
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.widget.Toast;
+package com.naughtybitch.discogsclient.sell;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
@@ -19,10 +11,29 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
-import com.google.android.material.navigation.NavigationView;
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+import android.widget.Toast;
 
-public class ProfileActivity extends AppCompatActivity implements
-        ProfileFragment.OnFragmentInteractionListener {
+import com.google.android.material.appbar.CollapsingToolbarLayout;
+import com.google.android.material.navigation.NavigationView;
+import com.naughtybitch.discogsclient.explore.ExploreActivity;
+import com.naughtybitch.discogsclient.MainActivity;
+import com.naughtybitch.discogsclient.profile.ProfileActivity;
+import com.naughtybitch.discogsclient.R;
+import com.naughtybitch.discogsclient.settings.SettingsActivity;
+import com.naughtybitch.discogsclient.buy.BuyMusicActivity;
+import com.naughtybitch.discogsclient.wishlist.WishlistActivity;
+
+public class SellMusicActivity extends AppCompatActivity implements
+        SellMusicFragment.OnFragmentInteractionListener,
+        OrderFragment.OnFragmentInteractionListener,
+        OfferFragment.OnFragmentInteractionListener,
+        InventoryFragment.OnFragmentInteractionListener {
 
     private DrawerLayout drawerLayout;
     private ActionBarDrawerToggle toggle;
@@ -31,10 +42,12 @@ public class ProfileActivity extends AppCompatActivity implements
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile);
+        setContentView(R.layout.activity_sell_music);
 
         // Custom ActionBar
         final Toolbar myToolbar = (Toolbar) findViewById(R.id.my_toolbar);
+        final CollapsingToolbarLayout collapsingToolbarLayout = (CollapsingToolbarLayout) findViewById(R.id.collapse_toolbar_layout);
+        collapsingToolbarLayout.setTitleEnabled(false);
         setSupportActionBar(myToolbar);
 
         // DrawerLayout
@@ -46,11 +59,25 @@ public class ProfileActivity extends AppCompatActivity implements
         if (actionBar != null) {
             actionBar.setHomeButtonEnabled(true);
             actionBar.setDisplayHomeAsUpEnabled(true);
-            actionBar.setDisplayShowTitleEnabled(false);
         }
+        getSupportFragmentManager().addOnBackStackChangedListener(
+                new FragmentManager.OnBackStackChangedListener() {
+                    @Override
+                    public void onBackStackChanged() {
+                        if (getSupportFragmentManager().getBackStackEntryCount() == 0) {
+                            toggle.setDrawerIndicatorEnabled(true);
+                            setTitle(R.string.music_sell);
+                        }
+                        if (getSupportFragmentManager().getBackStackEntryCount() > 0) {
+                            toggle.setDrawerIndicatorEnabled(false);
+                            getSupportActionBar().setHomeButtonEnabled(true);
+                            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+                        }
+                    }
+                });
 
         // Create a new Fragment to be placed in the activity layout
-        ProfileFragment firstFragment = ProfileFragment.newInstance();
+        SellMusicFragment firstFragment = SellMusicFragment.newInstance();
 
         // In case this activity was started with special instructions from an
         // Intent, pass the Intent's extras to the fragment as arguments
@@ -65,36 +92,37 @@ public class ProfileActivity extends AppCompatActivity implements
     public void navigationViewHandler() {
         navigationView = (NavigationView) findViewById(R.id.navigation_view);
         navigationView.setItemIconTintList(null);
-        navigationView.setCheckedItem(R.id.profile);
+        navigationView.setCheckedItem(R.id.sell_music);
         navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
                 switch (menuItem.getItemId()) {
                     case R.id.home:
-                        startActivity(new Intent(ProfileActivity.this, MainActivity.class));
+                        startActivity(new Intent(SellMusicActivity.this, MainActivity.class));
                         break;
                     case R.id.profile:
-                        drawerLayout.closeDrawer(GravityCompat.START);
+                        startActivity(new Intent(SellMusicActivity.this, ProfileActivity.class));
+                        Toast.makeText(SellMusicActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.wish_list:
-                        startActivity(new Intent(ProfileActivity.this, WishlistActivity.class));
-                        Toast.makeText(ProfileActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(SellMusicActivity.this, WishlistActivity.class));
+                        Toast.makeText(SellMusicActivity.this, "ProfileFragment", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.buy_music:
-                        startActivity(new Intent(ProfileActivity.this, BuyMusicActivity.class));
-                        Toast.makeText(ProfileActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(SellMusicActivity.this, BuyMusicActivity.class));
+                        Toast.makeText(SellMusicActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.sell_music:
-                        startActivity(new Intent(ProfileActivity.this, SellMusicActivity.class));
-                        Toast.makeText(ProfileActivity.this, "SellMusicActivity", Toast.LENGTH_SHORT).show();
+                        drawerLayout.closeDrawer(GravityCompat.START);
+                        Toast.makeText(SellMusicActivity.this, "BuyMusicActivity", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.settings:
-                        startActivity(new Intent(ProfileActivity.this, SettingsActivity.class));
-                        Toast.makeText(ProfileActivity.this, "SettingsActivity", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(SellMusicActivity.this, SettingsActivity.class));
+                        Toast.makeText(SellMusicActivity.this, "SettingsActivity", Toast.LENGTH_SHORT).show();
                         break;
                     case R.id.explore:
-                        startActivity(new Intent(ProfileActivity.this, ExploreActivity.class));
-                        Toast.makeText(ProfileActivity.this, "ExploreActivity", Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(SellMusicActivity.this, ExploreActivity.class));
+                        Toast.makeText(SellMusicActivity.this, "ExploreActivity", Toast.LENGTH_SHORT).show();
                         break;
                     default:
                         return true;
@@ -128,8 +156,15 @@ public class ProfileActivity extends AppCompatActivity implements
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
 
         switch (item.getItemId()) {
+            case android.R.id.home:
+                FragmentManager fm = getSupportFragmentManager();
+                if (fm.getBackStackEntryCount() > 0) {
+                    fm.popBackStack();
+                    return true;
+                }
+                break;
             case R.id.search:
-                Intent intent = new Intent(ProfileActivity.this, ExploreActivity.class);
+                Intent intent = new Intent(SellMusicActivity.this, ExploreActivity.class);
                 startActivity(intent);
                 break;
         }
