+++ src/main/java/com/naughtybitch/discogsclient/sell/SellMusicFragment.java
@@ -1,4 +1,4 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.sell;
 
 import android.content.Context;
 import android.net.Uri;
@@ -7,20 +7,31 @@ import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.Toast;
+
+import com.naughtybitch.discogsclient.R;
+import com.naughtybitch.discogsclient.buy.BuyMusicFragment;
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link OfferFragment.OnFragmentInteractionListener} interface
+ * {@link BuyMusicFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link OfferFragment#newInstance} factory method to
+ * Use the {@link BuyMusicFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class OfferFragment extends Fragment {
+public class SellMusicFragment extends Fragment implements
+        View.OnClickListener,
+        OrderFragment.OnFragmentInteractionListener,
+        OfferFragment.OnFragmentInteractionListener,
+        InventoryFragment.OnFragmentInteractionListener {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -32,16 +43,48 @@ public class OfferFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public OfferFragment() {
+    public SellMusicFragment() {
         // Required empty public constructor
     }
 
-    public static OfferFragment newInstance() {
-        Bundle args = new Bundle();
-        OfferFragment fragment = new OfferFragment();
-        fragment.setArguments(args);
-        return fragment;
+    private void buttonOnClickListener(View v) {
+        Button btn_order = v.findViewById(R.id.order);
+        Button btn_offer = v.findViewById(R.id.offer);
+        Button btn_inven = v.findViewById(R.id.inventory);
+        btn_order.setOnClickListener(this);
+        btn_offer.setOnClickListener(this);
+        btn_inven.setOnClickListener(this);
     }
+
+    @Override
+    public void onFragmentInteraction(Uri uri) {
+        // Do smt
+    }
+
+    public void onClick(View v) {
+        switch (v.getId()) {
+            case R.id.order:
+                Toast.makeText(getActivity(), "????????", Toast.LENGTH_SHORT).show();
+                navigateToFragment(OrderFragment.newInstance());
+                break;
+            case R.id.offer:
+                navigateToFragment(OfferFragment.newInstance());
+                break;
+            case R.id.inventory:
+                navigateToFragment(InventoryFragment.newInstance());
+                break;
+        }
+    }
+
+    private void navigateToFragment(Fragment fragment) {
+        FragmentManager fm = getActivity().getSupportFragmentManager();
+        FragmentTransaction fragmentTransaction = fm.beginTransaction();
+        fragmentTransaction.replace(R.id.fragment_container, fragment);
+        fragmentTransaction.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
+        fragmentTransaction.addToBackStack(null);
+        fragmentTransaction.commit();
+    }
+
 
     /**
      * Use this factory method to create a new instance of
@@ -49,14 +92,21 @@ public class OfferFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment OfferFragment.
+     * @return A new instance of fragment BuyMusicFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static OfferFragment newInstance(String param1, String param2) {
-        OfferFragment fragment = new OfferFragment();
+    public static SellMusicFragment newInstance(String param1, String param2) {
+        SellMusicFragment fragment = new SellMusicFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    public static SellMusicFragment newInstance() {
+        Bundle args = new Bundle();
+        SellMusicFragment fragment = new SellMusicFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -74,7 +124,8 @@ public class OfferFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.fragment_offer, container, false);
+        View v = inflater.inflate(R.layout.fragment_sell_music, container, false);
+        buttonOnClickListener(v);
         return v;
     }
