+++ src/main/java/com/naughtybitch/discogsclient/wishlist/WishlistFragment.java
@@ -1,27 +1,27 @@
-package com.naughtybitch.discogsclient;
+package com.naughtybitch.discogsclient.wishlist;
 
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+
+import com.naughtybitch.discogsclient.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link PurchasesFragment.OnFragmentInteractionListener} interface
+ * {@link WishlistFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link PurchasesFragment#newInstance} factory method to
+ * Use the {@link WishlistFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class PurchasesFragment extends Fragment {
+public class WishlistFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -33,8 +33,15 @@ public class PurchasesFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public PurchasesFragment() {
+    public WishlistFragment() {
         // Required empty public constructor
+    }
+
+    public static WishlistFragment newInstance() {
+        Bundle args = new Bundle();
+        WishlistFragment fragment = new WishlistFragment();
+        fragment.setArguments(args);
+        return fragment;
     }
 
     /**
@@ -43,21 +50,14 @@ public class PurchasesFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment PurchasesFragment.
+     * @return A new instance of fragment WishlistFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static PurchasesFragment newInstance(String param1, String param2) {
-        PurchasesFragment fragment = new PurchasesFragment();
+    public static WishlistFragment newInstance(String param1, String param2) {
+        WishlistFragment fragment = new WishlistFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public static PurchasesFragment newInstance() {
-        Bundle args = new Bundle();
-        PurchasesFragment fragment = new PurchasesFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -71,12 +71,11 @@ public class PurchasesFragment extends Fragment {
         }
     }
 
-    @Nullable
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.fragment_purchases, container, false);
-        return v;
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_wishlist, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
