+++ src/main/res/layout/activity_settings.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:clipToPadding="true"
     android:fitsSystemWindows="true"
-    tools:context=".SettingsActivity">
+    tools:context=".settings.SettingsActivity">
 
     <include layout="@layout/app_bar_main">
