+++ src/main/res/layout/activity_wishlist.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:clipToPadding="true"
     android:fitsSystemWindows="true"
-    tools:context=".WishlistActivity">
+    tools:context=".wishlist.WishlistActivity">
 
     <include
         layout="@layout/app_bar_main"
