+++ src/main/res/layout/card_row.xml
@@ -3,41 +3,41 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:paddingTop="10dp"
-    android:paddingBottom="10dp">
+    android:layout_marginBottom="10dp"
+    android:orientation="vertical">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        app:cardCornerRadius="4dp">
+        android:background="@color/light_black"
+        android:backgroundTint="@color/light_black"
+        app:cardCornerRadius="0dp">
 
-        <LinearLayout
+        <RelativeLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
+            android:layout_height="wrap_content">
 
-            <de.hdodenhof.circleimageview.CircleImageView
+            <ImageView
                 android:id="@+id/card_image"
-                android:layout_width="96dp"
-                android:layout_height="96dp"
-                android:src="@drawable/lemons"
-                app:civ_border_color="@color/light_black"
-                app:civ_border_width="2dp">
+                android:layout_width="100dp"
+                android:layout_height="100dp"
+                android:contentDescription="@string/lemon">
 
-            </de.hdodenhof.circleimageview.CircleImageView>
+            </ImageView>
 
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
+                android:layout_toEndOf="@id/card_image"
                 android:orientation="vertical">
 
                 <TextView
                     android:id="@+id/card_title"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:text="@string/title">
+                    android:text="@string/title"
+                    android:textColor="@color/white">
 
                 </TextView>
 
@@ -45,13 +45,14 @@
                     android:id="@+id/card_genre"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
-                    android:text="@string/genre">
+                    android:text="@string/genre"
+                    android:textColor="@color/white">
 
                 </TextView>
             </LinearLayout>
 
 
-        </LinearLayout>
+        </RelativeLayout>
 
     </androidx.cardview.widget.CardView>
