+++ src/main/res/layout/fragment_buy_music.xml
@@ -7,7 +7,7 @@
     android:background="@drawable/gradient_color_market_buy"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".BuyMusicFragment">
+    tools:context=".buy.BuyMusicFragment">
 
     <LinearLayout
         style="?android:attr/buttonBarStyle"
