+++ src/main/res/layout/fragment_cart.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".CartFragment">
+    tools:context=".buy.CartFragment">
 
     <ImageView
         android:layout_width="132dp"
