+++ src/main/res/layout/fragment_dw.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".DWFragment">
+    tools:context=".buy.DWFragment">
 
     <ImageView
         android:layout_width="90dp"
