+++ src/main/res/layout/fragment_inventory.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".InventoryFragment">
+    tools:context=".sell.InventoryFragment">
 
     <ImageView
         android:layout_width="90dp"
