+++ src/main/res/layout/fragment_offer.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".OfferFragment">
+    tools:context=".sell.OfferFragment">
 
     <ImageView
         android:layout_width="90dp"
