+++ src/main/res/layout/fragment_oim.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".OIMFragment">
+    tools:context=".buy.OIMFragment">
 
     <ImageView
         android:layout_width="90dp"
