+++ src/main/res/layout/fragment_order.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".OrderFragment">
+    tools:context=".sell.OrderFragment">
 
     <ImageView
         android:layout_width="90dp"
