+++ src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ProfileFragment">
+    tools:context=".profile.ProfileFragment">
 
 <!--    <ImageView-->
 <!--        android:layout_width="wrap_content"-->
