+++ src/main/res/layout/fragment_purchases.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".PurchasesFragment">
+    tools:context=".buy.PurchasesFragment">
 
     <ImageView
         android:layout_width="90dp"
