+++ src/main/res/layout/fragment_suggestions.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/delicate"
-    tools:context=".SuggestionsFragment">
+    tools:context=".album.SuggestionsFragment">
 
     <include
         layout="@layout/suggestion_items"
