+++ src/main/res/xml/account_settings_preferences.xml
@@ -58,7 +58,7 @@
 
     <Preference
         app:allowDividerAbove="true"
-        app:fragment="com.naughtybitch.discogsclient.SettingsActivity$BuyingShippingAddressFragment"
+        app:fragment="com.naughtybitch.discogsclient.settings.SettingsActivity$BuyingShippingAddressFragment"
         app:iconSpaceReserved="false"
         app:key="buy_ship_address"
         app:title="@string/buying_ship_address" />
